put OrganizationCombinedRuleIndexEndpoint tests in their own file
